Update README.md
Conclusion:
By the end of this project, we aim to have a robust model that can predict the success of movies based on their attributes. This model can be used by movie production companies to make informed decisions about their projects, ultimately leading to better financial outcomes and audience satisfaction.